feat: CLIN-1046 - refractoring
