+++ src/main/java/bio/ferlab/clin/es/builder/PrescriptionDataBuilder.java
@@ -1,14 +1,8 @@
 package bio.ferlab.clin.es.builder;
 
 import bio.ferlab.clin.es.config.ResourceDaoConfiguration;
-import bio.ferlab.clin.es.data.FamilyGroupInfoData;
-import bio.ferlab.clin.es.data.PatientData;
 import bio.ferlab.clin.es.data.PrescriptionData;
-import bio.ferlab.clin.utils.Extensions;
-import ca.uhn.fhir.rest.api.server.IBundleProvider;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
-import org.hl7.fhir.instance.model.api.IBaseCoding;
-import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,16 +11,11 @@ import org.springframework.stereotype.Component;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 @Component
 public class PrescriptionDataBuilder {
     private static final Logger log = LoggerFactory.getLogger(PrescriptionDataBuilder.class);
-
-    private static final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
-    public static final String MRN_CODE = "MR";
 
     private final ResourceDaoConfiguration configuration;
     private final CommonDataBuilder commonDataBuilder;
@@ -36,188 +25,34 @@ public class PrescriptionDataBuilder {
         this.commonDataBuilder = commonDataBuilder;
     }
 
-    public List<PrescriptionData> fromIds(Set<String> ids, RequestDetails requestDetails) {
+    public List<PrescriptionData> fromIds(Set<String> ids, RequestDetails requestDetails, PrescriptionDataType type) {
         final List<PrescriptionData> prescriptionDataList = new ArrayList<>();
         for (final String serviceRequestId : ids) {
             final PrescriptionData prescriptionData = new PrescriptionData();
             final ServiceRequest serviceRequest = this.configuration.serviceRequestDAO.read(new IdType(serviceRequestId), requestDetails);
-            if (!serviceRequest.hasSubject()) {
-                log.error(String.format("Cannot index ServiceRequest [%s]: resource has no subject.", serviceRequestId));
-                continue;
-            }
-            final Reference subject = serviceRequest.getSubject();
-            final Patient patient = this.configuration.patientDAO.read(new IdType(subject.getReference()), requestDetails);
+            if (isOfType(serviceRequest, type)) {
+                if (!serviceRequest.hasSubject()) {
+                    log.error(String.format("Cannot index ServiceRequest [%s]: resource has no subject.", serviceRequestId));
+                    continue;
+                }
+                final Reference subject = serviceRequest.getSubject();
+                final Patient patient = this.configuration.patientDAO.read(new IdType(subject.getReference()), requestDetails);
 
-            this.handlePatient(patient, prescriptionData.getPatientInfo());
-            this.handleServiceRequest(serviceRequest, prescriptionData);
-            if (patient.hasExtension(Extensions.FAMILY_ID)) {
-                final Extension extension = patient.getExtensionByUrl(Extensions.FAMILY_ID);
-                final Reference ref = (Reference) extension.getValue();
-                final Group group = this.configuration.groupDao.read(ref.getReferenceElement());
-                this.handleFamilyGroup(group, prescriptionData.getFamilyInfo());
+                this.commonDataBuilder.handlePatient(patient, prescriptionData.getPatientInfo());
+                this.commonDataBuilder.finalize(prescriptionData.getPatientInfo());
+                
+                this.commonDataBuilder.handleServiceRequest(serviceRequest, prescriptionData);
+                this.commonDataBuilder.finalize(prescriptionData);
+                
+                prescriptionDataList.add(prescriptionData);
             }
-            // clean-up + full text
-            prescriptionData.getPatientInfo().setRequests(null);
-            prescriptionData.applyFullText();
-            prescriptionData.getSecurityTags().addAll(extractMetaTags(serviceRequest));
-            prescriptionDataList.add(prescriptionData);
         }
         return prescriptionDataList;
     }
 
-    void handlePatient(Patient patient, PatientData patientInfo) {
-        patientInfo.setCid(patient.getIdElement().getIdPart());
-        final List<String> mrns = Objects.requireNonNull(patient.getIdentifier()).stream()
-                .filter(id -> id.getType().getCodingFirstRep().getCode().contentEquals(MRN_CODE))
-                .map(Identifier::getValue).collect(Collectors.toList());
-
-        patientInfo.getMrn().addAll(mrns);
-        patientInfo.setGender(patient.getGender().getDisplay());
-        patientInfo.getSecurityTags().addAll(extractMetaTags(patient));
-
-        if (patient.hasName()) {
-            final Name name = extractName(patient.getName());
-            patientInfo.setLastName(name.lastName);
-            patientInfo.setFirstName(name.firstName);
-            patientInfo.setLastNameFirstName(name.lastNameFirstName);
-        }
-        if (patient.getIdentifier().size() > 1) {
-            patientInfo.setRamq(patient.getIdentifier().get(1).getValue());
-        }
-
-        if (patient.hasExtension(Extensions.IS_PROBAND)) {
-            final Extension extension = patient.getExtensionByUrl(Extensions.IS_PROBAND);
-            if (extension.hasValue()) {
-                final BooleanType value = (BooleanType) extension.getValue();
-                patientInfo.setPosition(value.booleanValue() ? "Proband" : "Parent");
-            }
-        }
-
-        if (patient.hasExtension(Extensions.IS_FETUS)) {
-            final Extension extension = patient.getExtensionByUrl(Extensions.IS_FETUS);
-            if (extension.hasValue()) {
-                final BooleanType value = (BooleanType) extension.getValue();
-                patientInfo.setFetus(value.booleanValue());
-            }
-        }
-
-        if (patient.hasManagingOrganization()) {
-            final String id = patient.getManagingOrganization().getReference();
-            final Organization organization = configuration.organizationDAO.read(new IdType(id));
-            patientInfo.getOrganization().setCid(organization.getIdElement().getIdPart());
-            patientInfo.getOrganization().setName(organization.hasName() ? organization.getName() : "");
-        }
-
-        if (patient.hasBirthDate()) {
-            patientInfo.setBirthDate(simpleDateFormat.format(patient.getBirthDate()));
-        }
-
+    private boolean isOfType(ServiceRequest serviceRequest, PrescriptionDataType type) {
+        return PrescriptionDataType.ANY.equals(type) 
+            || serviceRequest.getMeta().getProfile().stream().anyMatch(s -> type.value.equals(s.getValue()));
     }
 
-    void handleServiceRequest(ServiceRequest serviceRequest, PrescriptionData prescriptionData) {
-        prescriptionData.setCid(serviceRequest.getIdElement().getIdPart());
-        prescriptionData.setStatus(serviceRequest.getStatus().toCode());
-        if (serviceRequest.hasCode()) {
-            final CodeableConcept code = serviceRequest.getCode();
-            if (code.hasCoding()) {
-                final Coding coding = code.getCoding().get(0);
-                prescriptionData.getAnalysis().setCode(coding.getCode());
-                prescriptionData.getAnalysis().setDisplay(coding.getDisplay());
-            }
-        }
-
-        if (serviceRequest.hasExtension(Extensions.IS_SUBMITTED)) {
-            final Extension extension = serviceRequest.getExtensionByUrl(Extensions.IS_SUBMITTED);
-            final BooleanType valueBoolean = (BooleanType) extension.getValue();
-            prescriptionData.setSubmitted(valueBoolean.getValue());
-        }
-        
-        if (serviceRequest.hasExtension(Extensions.PROCEDURE_DIRECTED_BY)) {
-            final Extension extension = serviceRequest.getExtensionByUrl(Extensions.PROCEDURE_DIRECTED_BY);
-            final Reference ref = (Reference) extension.getValue();
-            final PractitionerRole role = this.configuration.practitionerRoleDao.read(ref.getReferenceElement());
-            final Practitioner practitioner = this.configuration.practitionerDao.read(role.getPractitioner().getReferenceElement());
-            final Name name = extractName(practitioner.getName());
-            prescriptionData.getApprover().setCid(practitioner.getIdElement().getIdPart());
-            prescriptionData.getApprover().setFirstName(name.firstName);
-            prescriptionData.getApprover().setLastName(name.lastName);
-            prescriptionData.getApprover().setLastNameFirstName(name.lastNameFirstName);
-        }
-        
-        if (serviceRequest.hasPerformer()) {
-            prescriptionData.setLaboratory(serviceRequest.getPerformer().get(0).getReference());
-        }
-
-        if (serviceRequest.hasAuthoredOn()) {
-            prescriptionData.setAuthoredOn(simpleDateFormat.format(serviceRequest.getAuthoredOn()));
-        }
-
-        if (serviceRequest.hasRequester()) {
-            final String id = serviceRequest.getRequester().getReference();
-            final Practitioner practitioner = configuration.practitionerDao.read(new IdType(id));
-            final Name name = extractName(practitioner.getName());
-            prescriptionData.getPrescriber().setCid(practitioner.getIdElement().getIdPart());
-            prescriptionData.getPrescriber().setLastName(name.lastName);
-            prescriptionData.getPrescriber().setFirstName(name.firstName);
-            prescriptionData.getPrescriber().setLastNameFirstName(name.lastNameFirstName);
-        }
-
-        if (serviceRequest.hasIdentifier()) {
-            final var identifier = serviceRequest.getIdentifier().get(0);
-            if (identifier.hasValue()) {
-                prescriptionData.setMrn(identifier.getValue());
-            }
-            if (identifier.hasAssigner()) {
-                final String id = identifier.getAssigner().getReference();
-                final Organization organization = configuration.organizationDAO.read(new IdType(id));
-                prescriptionData.getOrganization().setCid(organization.getIdElement().getIdPart());
-                prescriptionData.getOrganization().setName(organization.hasName() ? organization.getName() : "");
-            }
-        }
-
-        prescriptionData.buildState();
-    }
-
-    void handleFamilyGroup(Group group, FamilyGroupInfoData familyGroupInfo) {
-        familyGroupInfo.setCid(group.getIdElement().getIdPart());
-        if (group.hasExtension(Extensions.FAMILY_ID)) {
-            final Extension extension = group.getExtensionByUrl(Extensions.FAMILY_ID);
-            final Coding coding = (Coding) extension.getValue();
-            familyGroupInfo.setType(coding.getDisplay());
-        } else {
-            familyGroupInfo.setType("solo");
-        }
-    }
-
-    private static Name extractName(List<HumanName> humanNames) {
-        final HumanName name = humanNames.get(0);
-        return new Name(name.getFamily(), name.getGivenAsSingleString());
-    }
-
-    @SuppressWarnings("unchecked")
-    private <T extends IBaseResource> List<T> getListFromProvider(IBundleProvider provider) {
-        final List<T> resources = new ArrayList<>();
-        if (!provider.isEmpty()) {
-            for (IBaseResource sr : provider.getResources(0, provider.size())) {
-                resources.add((T) sr);
-            }
-        }
-        return resources;
-    }
-
-    private List<String> extractMetaTags(IBaseResource resource) {
-        return resource.getMeta().getSecurity().stream().map(IBaseCoding::getCode).collect(Collectors.toList());
-    }
-
-    private static class Name {
-        public final String lastName;
-        public final String firstName;
-        public final String lastNameFirstName;
-
-        public Name(String lastName, String firstName) {
-            this.lastName = lastName;
-            this.firstName = firstName;
-            this.lastNameFirstName = lastName + ", " + firstName;
-        }
-    }
 }
