+++ src/main/java/bio/ferlab/clin/es/builder/PrescriptionDataType.java
@@ -1,7 +1,8 @@
-package bio.ferlab.clin.es.builder.nanuq;
+package bio.ferlab.clin.es.builder;
 
 public enum PrescriptionDataType {
 
+  ANY(null),
   ANALYSIS("http://fhir.cqgc.ferlab.bio/StructureDefinition/cqgc-analysis-request"),
   SEQUENCING("http://fhir.cqgc.ferlab.bio/StructureDefinition/cqgc-sequencing-request");
