+++ src/main/java/bio/ferlab/clin/es/indexer/LegacyNanuqIndexer.java
@@ -1,9 +1,9 @@
 package bio.ferlab.clin.es.indexer;
 
 import bio.ferlab.clin.es.ElasticsearchRestClient;
-import bio.ferlab.clin.es.builder.nanuq.PatientDataBuilder;
-import bio.ferlab.clin.es.builder.nanuq.PrescriptionDataBuilder;
-import bio.ferlab.clin.es.builder.nanuq.PrescriptionDataType;
+import bio.ferlab.clin.es.builder.PatientDataBuilder;
+import bio.ferlab.clin.es.builder.PrescriptionDataBuilder;
+import bio.ferlab.clin.es.builder.PrescriptionDataType;
 import bio.ferlab.clin.es.data.PatientData;
 import bio.ferlab.clin.es.data.PrescriptionData;
 import bio.ferlab.clin.es.extractor.PatientIdExtractor;
@@ -11,7 +11,6 @@ import bio.ferlab.clin.es.extractor.ServiceRequestIdExtractor;
 import bio.ferlab.clin.properties.BioProperties;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -20,7 +19,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Component
-public class NanuqIndexer extends Indexer {
+public class LegacyNanuqIndexer extends Indexer {
   
   private final ServiceRequestIdExtractor serviceRequestIdExtractor;
   private final PrescriptionDataBuilder prescriptionDataBuilder;
@@ -29,12 +28,12 @@ public class NanuqIndexer extends Indexer {
   private final BioProperties bioProperties;
   private final IndexerTools tools;
   
-  public NanuqIndexer(ServiceRequestIdExtractor serviceRequestIdExtractor,
-                      @Qualifier("NanuqPrescriptionDataBuilder") PrescriptionDataBuilder prescriptionDataBuilder,
-                      PatientIdExtractor patientIdExtractor,
-                      @Qualifier("NanuqPatientDataBuilder") PatientDataBuilder patientDataBuilder,
-                      BioProperties bioProperties,
-                      IndexerTools tools) {
+  public LegacyNanuqIndexer(ServiceRequestIdExtractor serviceRequestIdExtractor,
+                            PrescriptionDataBuilder prescriptionDataBuilder,
+                            PatientIdExtractor patientIdExtractor,
+                            PatientDataBuilder patientDataBuilder,
+                            BioProperties bioProperties,
+                            IndexerTools tools) {
     this.serviceRequestIdExtractor = serviceRequestIdExtractor;
     this.prescriptionDataBuilder = prescriptionDataBuilder;
     this.patientIdExtractor = patientIdExtractor;
@@ -47,33 +46,45 @@ public class NanuqIndexer extends Indexer {
   protected void doIndex(RequestDetails requestDetails, IBaseResource resource) {
     final Set<String> patientIds = patientIdExtractor.extract(resource);
     List<PatientData> patients = this.indexPatients(requestDetails, patientIds);
-    // re-index linked distinct prescriptions
-    final Set<String> linkedPrescriptions = patients.stream().flatMap(p -> p.getRequests().stream().map(PrescriptionData::getCid)).collect(Collectors.toSet());
-    this.indexPrescriptions(requestDetails, linkedPrescriptions);
     
     final Set<String> prescriptionIds = serviceRequestIdExtractor.extract(resource);
     List<PrescriptionData> prescriptions = this.indexPrescriptions(requestDetails, prescriptionIds);
-    final Set<String> linkedPatients = prescriptions.stream().map(p -> p.getPatientInfo().getCid()).collect(Collectors.toSet());
+
+    // re-index linked distinct prescriptions
+    final Set<String> linkedPrescriptions = patients.stream().flatMap(p -> p.getRequests().stream().map(PrescriptionData::getCid)).collect(Collectors.toSet());
+    linkedPrescriptions.removeAll(prescriptionIds); // ignore if indexed before
+    this.indexPrescriptions(requestDetails, linkedPrescriptions);
+
     // re-index linked distinct patients
+    final Set<String> linkedPatients = prescriptions.stream().map(p -> p.getPatientInfo().getCid()).collect(Collectors.toSet());
+    linkedPatients.removeAll(patientIds); // ignore if indexed before
     this.indexPatients(requestDetails, linkedPatients);
   }
   
   private List<PatientData> indexPatients(RequestDetails requestDetails, Set<String> ids) {
     final List<PatientData> patients = patientDataBuilder.fromIds(ids, requestDetails);
-    patients.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getNanuqEsPatientsIndex()));
+    patients.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getEsPatientsIndex()));
     return patients;
   }
 
   private List<PrescriptionData> indexPrescriptions(RequestDetails requestDetails, Set<String> ids) {
-    final List<PrescriptionData> analysis = prescriptionDataBuilder.fromIds(ids, requestDetails,
-        PrescriptionDataType.ANALYSIS);
-    analysis.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getNanuqEsAnalysesIndex()));
-    
-    final List<PrescriptionData> sequencing = prescriptionDataBuilder.fromIds(ids, requestDetails,
-        PrescriptionDataType.SEQUENCING);
-    sequencing.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getNanuqEsSequencingsIndex()));
-    
-    return Stream.concat(analysis.stream(), sequencing.stream()).collect(Collectors.toList());
+    if (bioProperties.isNanuqEnabled()) {
+      final List<PrescriptionData> analysis = prescriptionDataBuilder.fromIds(ids, requestDetails,
+          PrescriptionDataType.ANALYSIS);
+      analysis.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getEsAnalysesIndex()));
+
+      final List<PrescriptionData> sequencing = prescriptionDataBuilder.fromIds(ids, requestDetails,
+          PrescriptionDataType.SEQUENCING);
+      sequencing.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getEsPrescriptionsIndex()));
+
+      return Stream.concat(analysis.stream(), sequencing.stream()).collect(Collectors.toList());
+    } else {  // LEGACY
+      final List<PrescriptionData> prescriptions = prescriptionDataBuilder.fromIds(ids, requestDetails,
+          PrescriptionDataType.ANY);
+      prescriptions.forEach(e -> indexToEs(e.getCid(), e, bioProperties.getEsPrescriptionsIndex()));
+      
+      return prescriptions;
+    }
   }
   
   private void indexToEs(String id, Object document, String indexName) {
